[CORE-1703] leave formatter job to formatter; move json yml helpers to separate class; refactor a little custom socket formatter;
