+++ syncano_cli/custom_sockets/formatters.py
@@ -1,261 +1,18 @@
 # -*- coding: utf-8 -*-
-import os
 import sys
-from collections import defaultdict
 
 import click
 import six
 import yaml
 from syncano_cli.base.formatters import Formatter
 from syncano_cli.base.options import ColorSchema, DefaultOpt, SpacedOpt, TopSpacedOpt, WarningOpt
-from syncano_cli.custom_sockets.exceptions import BadYAMLDefinitionInEndpointsException
-from syncano_cli.sync.scripts import ALLOWED_RUNTIMES
-
-
-class DependencyTypeE():
-    CLASS = 'class'
-    SCRIPT = 'script'
 
 
 class SocketFormatter(Formatter):
 
-    SOCKET_FIELDS = ['name', 'description', 'endpoints', 'dependencies']
-    SOCKET_DISPLAY_FIELDS = ['name', 'description', 'status', 'status_info', 'endpoints']
-    HTTP_METHODS = ['GET', 'POST', 'DELETE', 'PUT', 'PATCH']
-    ENDPOINT_TYPES = ['script']
-    DEPENDENCY_TYPES = {'scripts': DependencyTypeE.SCRIPT, 'classes': DependencyTypeE.CLASS}
+    SOCKET_LIST_FIELDS = ['name', 'description', 'status', 'status_info', 'endpoints']
+    SOCKET_DETAILS_FIELDS = ['config', 'endpoints', 'dependencies', 'metadata']
     ENDPOINT_FIELDS = ['calls', 'acl', 'metadata']
-
-    @classmethod
-    def get_dependency_handlers(cls):
-        handlers = {}
-        for yml_name, dependency_type in six.iteritems(cls.DEPENDENCY_TYPES):
-            handlers[dependency_type] = getattr(cls, 'get_{}_dependency'.format(dependency_type))
-        return handlers
-
-    @classmethod
-    def to_yml(cls, socket_object):
-        """
-        A method which transform CustomSocket object from Syncano Python LIB into yml representation.
-        :param socket_object: the CustomSocket object;
-        :param scope: how many attributes should be processed and showed in output;
-        :return: the yml format of the CustomSocket;
-        """
-        yml_dict = {}
-        files = []
-
-        for socket_field in cls.SOCKET_FIELDS:
-            if socket_field == 'endpoints':
-                yml_dict[socket_field] = cls._yml_process_endpoints(socket_object.endpoints)
-            elif socket_field == 'dependencies':
-                yml_dict[socket_field], files = cls._yml_process_dependencies(socket_object.dependencies)
-            else:
-                # TODO: None for skip the description now;
-                yml_dict[socket_field] = getattr(socket_object, socket_field, None)
-
-        for meta_name, meta_data in six.iteritems(socket_object.metadata):
-            yml_dict[meta_name] = meta_data
-
-        return yaml.safe_dump(yml_dict, default_flow_style=False), files
-
-    @classmethod
-    def to_json(cls, socket_yml, directory=None):
-        """
-        A method which transform yml to the json. This json is used in API call (Syncano Python LIB)
-        :param socket_yml: the yml file path;
-        :return: the json format of the custom_socket;
-        """
-        metadata_dict = {}
-        api_dict = {}
-        for key, value in six.iteritems(socket_yml):
-            if key not in cls.SOCKET_FIELDS:
-                metadata_dict[key] = value
-            else:
-                if key == 'endpoints':
-                    api_dict[key] = cls._json_process_endpoints(endpoints=value)
-                elif key == 'dependencies':
-                    api_dict[key] = cls._json_process_dependencies(dependencies=value, directory=directory)
-                else:
-                    api_dict[key] = value
-
-        api_dict['metadata'] = metadata_dict
-        return api_dict
-
-    @classmethod
-    def _json_process_endpoints(cls, endpoints):
-        api_endpoints = {}
-
-        for name, endpoint_data in six.iteritems(endpoints):
-            api_endpoints[name] = {'calls': cls._get_calls(endpoint_data)}
-            api_endpoints[name].update({'metadata': cls._get_metadata(endpoint_data)})
-
-        return api_endpoints
-
-    @classmethod
-    def _get_metadata(cls, endpoint_data):
-        metadata = defaultdict(dict)
-        for data_key, data in six.iteritems(endpoint_data):
-            if data_key in cls.HTTP_METHODS:
-                for metadata_key, inner_data in six.iteritems(data):
-                    if metadata_key in cls.ENDPOINT_TYPES:
-                        continue
-                    if metadata_key == 'parameters':
-                        metadata[metadata_key][data_key] = inner_data
-                    else:
-                        metadata[metadata_key] = inner_data
-
-            elif data_key not in cls.ENDPOINT_TYPES:
-                if data_key == 'parameters':
-                    metadata[data_key]['*'] = data
-                else:
-                    metadata[data_key] = data
-                metadata[data_key] = data
-
-        return metadata
-
-    @classmethod
-    def _get_calls(cls, endpoint_data):
-        calls = []
-        keys = set(endpoint_data.keys())
-        if keys.intersection(set(cls.HTTP_METHODS)) and keys.intersection(set(cls.ENDPOINT_TYPES)):
-            raise BadYAMLDefinitionInEndpointsException()
-
-        for type_or_method in keys:
-            data = endpoint_data[type_or_method]
-
-            if type_or_method in cls.ENDPOINT_TYPES:
-                calls.append({
-                    'type': type_or_method,
-                    'methods': ['*'],
-                    'name': data,
-                })
-
-            elif type_or_method in cls.HTTP_METHODS:
-                for call_type, name in six.iteritems(data):
-
-                    if call_type in cls.ENDPOINT_TYPES:
-                        calls.append({
-                            'type': call_type,
-                            'methods': [type_or_method],
-                            'name': name
-                        })
-        return calls
-
-    @classmethod
-    def _json_process_dependencies(cls, dependencies, directory):
-        api_dependencies = []
-        dependency_handlers = cls.get_dependency_handlers()
-        for dependencies_type, dependency in six.iteritems(dependencies):
-            if dependencies_type in cls.DEPENDENCY_TYPES:
-                for dependency_name, data in six.iteritems(dependency):
-                    dependency_type = cls.DEPENDENCY_TYPES[dependencies_type]
-                    base_dependency_data = {
-                        'name': dependency_name,
-                        'type': dependency_type
-                    }
-                    typed_dependency_data = dependency_handlers[dependency_type](data, directory=directory)
-                    typed_dependency_data.update(base_dependency_data)
-                    api_dependencies.append(typed_dependency_data)
-        return api_dependencies
-
-    @classmethod
-    def get_script_dependency(cls, data, **kwargs):
-        """
-        Note: when definig new depenency processors use following pattern:
-        get_{name}_dependency -> where {name} is one of the defined in DepedencyTypeE
-        this allows to easily extend dependency handling;
-        :param data:
-        :param directory:
-        :return:
-        """
-        directory = kwargs.get('directory')
-        return {
-            'runtime_name': data['runtime_name'],
-            'source': cls._get_source(data['file'], directory),
-        }
-
-    @classmethod
-    def get_class_dependency(cls, data, **kwargs):
-        return {
-            'schema': data['schema']
-        }
-
-    @classmethod
-    def _get_source(cls, file_name, directory):
-        with open(os.path.join(directory, '{}'.format(file_name)), 'r+') as source_file:
-            return source_file.read()
-
-    @classmethod
-    def _yml_process_endpoints(cls, endpoints):
-        yml_endpoints = {}
-        for endpoint_name, endpoint_data in six.iteritems(endpoints):
-            yml_endpoints[endpoint_name] = cls._yml_process_calls(endpoint_data['calls'])
-            yml_endpoints.update(cls._yml_process_metadata(endpoint_data))
-        return yml_endpoints
-
-    @classmethod
-    def _yml_process_metadata(cls, endpoint_data):
-        return endpoint_data.get('metadata', {})  # some old Sockets do not have this field;
-
-    @classmethod
-    def _yml_process_calls(cls, data_calls):
-        calls = {}
-        for call in data_calls:
-            if call['methods'] == ['*']:
-                calls = {call['type']: call['name']}
-            else:
-                for method in call['methods']:
-                    calls.update({method: {call['type']: call['name']}})
-        return calls
-
-    @classmethod
-    def _yml_process_dependencies(cls, dependencies):
-        yml_dependencies = {}
-        scripts = {}
-        files = []
-
-        for dependency in dependencies:
-            if dependency['type'] == 'script':
-                file_name = '{name}{ext}'.format(name=dependency['name'],
-                                                 ext=ALLOWED_RUNTIMES[dependency['runtime_name']])
-
-                scripts[dependency['name']] = {
-                    'runtime_name': dependency['runtime_name'],
-                    'file': file_name
-                }
-                # create file list;
-                files.append(
-                    {
-                        'source': dependency['source'],
-                        'file_name': file_name
-                    }
-                )
-
-        yml_dependencies['scripts'] = scripts
-
-        return yml_dependencies, files
-
-    def display_socket_details(self, custom_socket, api_key):
-        self.write_block(
-            ['Details for Socket `{}` in `{}` instance.'.format(custom_socket.name,
-                                                                custom_socket.instance_name)],
-            SpacedOpt())
-        lines = self._display_list_details(custom_socket)
-        self.write_block(lines)
-        self.write('Config')
-        self.display_config(custom_socket.config)
-        self.separator()
-        self.write('Endpoints')
-        self._display_endpoints(custom_socket.endpoints, base_link=custom_socket.links.links_dict['endpoints'],
-                                api_key=api_key)
-        self.separator()
-        self.write('Dependencies')
-        self.format_list(custom_socket.dependencies, indent=1, skip_fields=['source'])
-        self.separator()
-        self.write('Metadata')
-        self.format_object(custom_socket.metadata, indent=1)
-        self.separator()
-        self.empty_line()
 
     def display_socket_list(self, socket_list, instance_name):
         if not socket_list:
@@ -266,12 +23,66 @@ class SocketFormatter(Formatter):
 
         for cs in socket_list:
             lines = self._display_list_details(custom_socket=cs)
-            lines.append('See: `syncano sockets details {}` for details.'.format(cs.name))
+            lines.append('See: `{}` {}'.format(
+                click.style('syncano sockets details {}'.format(cs.name), fg=ColorSchema.WARNING),
+                click.style('for details.', fg=ColorSchema.INFO)
+            ))
             self.write_block(lines, DefaultOpt(indent=2))
+
+    @classmethod
+    def format_endpoints_list(cls, socket_endpoints):
+        yml_dict = {'endpoints': []}
+        for endpoint in socket_endpoints:
+            endpoint_data = {'name': endpoint.name, 'path': endpoint.links.self}
+            endpoint_data.update({'methods': endpoint.allowed_methods})
+            yml_dict['endpoints'].append({'endpoint': endpoint_data})
+        return yaml.safe_dump(yml_dict, default_flow_style=False)
+
+    def display_socket_details(self, custom_socket, api_key):
+        self.write_block(
+            [
+                'Details for Socket `{}` in `{}` instance.'.format(
+                    custom_socket.name,
+                    custom_socket.instance_name
+                )
+            ],
+            SpacedOpt()
+        )
+
+        lines = self._display_list_details(custom_socket)
+        self.write_block(lines)
+
+        for socket_field in self.SOCKET_DETAILS_FIELDS:
+            self.display_block(
+                socket_field,
+                custom_socket,
+                base_link=custom_socket.links.links_dict['endpoints'],
+                api_key=api_key,
+                indent=1,
+                skip_fields=['source']
+            )
+
+    def display_block(self, block_name, socket, **kwargs):
+        handler, kw_names = self.display_handlers[block_name]
+        self.write(block_name.capitalize())
+        socket_data = getattr(socket, block_name)
+        new_kwargs = {key: value for key, value in six.iteritems(kwargs) if key in kw_names}
+        handler(socket_data, **new_kwargs)
+        self.separator()
+
+    @property
+    def display_handlers(self):
+        handlers = [
+            (self.display_config, ()),
+            (self.display_endpoints, ('base_link', 'api_key')),
+            (self.format_list, ('indent', 'skip_fields')),
+            (self.format_object, 'indent')
+        ]
+        return dict(zip(self.SOCKET_DETAILS_FIELDS, handlers))
 
     def _display_list_details(self, custom_socket):
         lines = []
-        for field in self.SOCKET_DISPLAY_FIELDS:
+        for field in self.SOCKET_LIST_FIELDS:
             value = getattr(custom_socket, field)
             if field == 'endpoints':
                 value = ', '.join(value.keys())
@@ -280,32 +91,42 @@ class SocketFormatter(Formatter):
             lines.append("{}: {}".format(field.capitalize().replace('_', ' '), value))
         return lines
 
-    def _display_endpoints(self, endpoints, base_link, api_key):
-        instance_part, endpoints_part = base_link.split('endpoints')
+    def display_endpoints(self, endpoints, base_link, api_key):
+
         for endpoint_name, endpoint_data in six.iteritems(endpoints):
             self.write('{e_name}:'.format(e_name=endpoint_name, ), DefaultOpt(indent=2), WarningOpt(), TopSpacedOpt())
             self.write('{url_label}: {link}'.format(
                 url_label=click.style('URL', fg=ColorSchema.PROMPT),
-                link=click.style(
-                    """{host}{instance_part}
-                    {endpoints_part}{name}/
-                    ?api_key={key}""".format(
-                        host='https://api.syncano.io',
-                        instance_part=instance_part,
-                        endpoints_part='endpoints{}'.format(endpoints_part),
-                        name=endpoint_name,
-                        key=api_key
-                    ), fg=ColorSchema.INFO
-                )
+                link=self._format_link(base_link, endpoint_name, api_key)
             ), DefaultOpt(indent=3))
-            for field in self.ENDPOINT_FIELDS:
-                handler_name = '_display_{}'.format(field)
-                data = endpoint_data.get(field)
-                try:
-                    handler = getattr(self, handler_name)
-                except AttributeError:
-                    continue
-                handler(data, indent=3)
+            self._display_endpoints_fields(endpoint_data)
+
+    @classmethod
+    def _format_link(cls, base_link, endpoint_name, api_key):
+        instance_part, endpoints_part = base_link.split('endpoints')
+        return '{link}'.format(
+            link=click.style(
+                """{host}{instance_part}
+                {endpoints_part}{name}/
+                ?api_key={key}""".format(
+                    host='https://api.syncano.io',
+                    instance_part=instance_part,
+                    endpoints_part='endpoints{}'.format(endpoints_part),
+                    name=endpoint_name,
+                    key=api_key
+                ), fg=ColorSchema.INFO
+            )
+        )
+
+    def _display_endpoints_fields(self, endpoint_data):
+        for field in self.ENDPOINT_FIELDS:
+            handler_name = '_display_{}'.format(field)
+            data = endpoint_data.get(field)
+            try:
+                handler = getattr(self, handler_name)
+            except AttributeError:
+                continue
+            handler(data, indent=3)
 
     def _display_calls(self, data, indent):
         if not data:
@@ -334,12 +155,3 @@ class SocketFormatter(Formatter):
             return
         self.write(click.style('Metadata:', fg=ColorSchema.PROMPT), DefaultOpt(indent=indent))
         self.format_object(data, DefaultOpt(indent=indent))
-
-    @classmethod
-    def format_endpoints_list(cls, socket_endpoints):
-        yml_dict = {'endpoints': []}
-        for endpoint in socket_endpoints:
-            endpoint_data = {'name': endpoint.name, 'path': endpoint.links.self}
-            endpoint_data.update({'methods': endpoint.allowed_methods})
-            yml_dict['endpoints'].append({'endpoint': endpoint_data})
-        return yaml.safe_dump(yml_dict, default_flow_style=False)
